requirements.txt
flask
flask-cors
requests
tzdata
user-agents
pandas
openpyxl
tika
python-dotenv
xmltodict
feedparser
google-auth 
google-auth-oauthlib 
google-auth-httplib2 
python-dotenv
google-analytics-data
google-api-python-client
google-cloud-storage
beautifulsoup4
lxml
pytrends==4.9.2
urllib3==2.0.7
psycopg2-binary 
sqlalchemy 
alembic
apscheduler
mysql-connector-python
psycopg2-binary
snowflake-connector-python
google-cloud-bigquery